Include value of unexpected Content-Type headers
The previous behavior just said there was an unexpected value, but
didn't tell you what it was. Including the received value makes it
easier to debug and fix the actual discrepancy.

I also changed "Content-type" to "Content-Type" (uppercase T) in these
error messages to match the canonical form of the header.