Allow to read Response single file definition
This is a proof-of-concept of reading definition from files which only
contains the definition of a single component.

The proof of concept only works for a response object reference and it
isn't maybe the best way of endowing the implementation with this
feature in how it has been implemented, however this is the solution
that I found in a few hours and mostly is for requesting feedback in how
to tackle the implementation of the feature not just for only responses
objects, but for any OpenAPI element which can be referenced.