Add content parameter validation
Fix crash in validation of parameters which are specified with
content rather than schema. The problem was that the presence
of a parameter with a nil schema in the Swagger spec would
crash the loop which validated over all parameters.

Added a callback to the request validation input so the user
can decode custom content types, and also implemented an
alternate code path for validation of content arguments. Added
an internal default parameter validator which is capable of
validating JSON encoded parameters.