fix: handle security requirements as OpenAPI 3 specifies (I think!!!)
From https://swagger.io/specification/#security-requirement-object:

> Security Requirement Objects that contain multiple schemes require
> that all schemes MUST be satisfied for a request to be authorized.
> This enables support for scenarios where multiple query parameters or
> HTTP headers are required to convey security information.

I'm fairly sure there was a bug in ValidateSecurityRequirements(),
which I've fixed. And I added a test case to TestValidateRequest() to
show the problem and prove that I've fixed it.

Unfortunately, this also breaks TestAnySecurityRequirementMet(), and
I'm not yet sure why. Not sure if the test is incorrect or if my fix
is incorrect.