Add openapi3.Schema.PropertyKeys for deterministic order
This commit adds the `PropertyKeys` property to the type
`openapi3.Schema` which contains the keys of the `Properties` map in the
order that they appear in the original YAML file. This is useful to
guarantee deterministic code generation.

This is done via a temporary fork of the YAML-to-JSON transformation library.
It will not be ready until invopop/yaml#13 is merged.