openapi3: process discriminator mapping values as refs (#1022)
* openapi3: make StringMap generic over value

This allows us to support more use cases than just a string, allowing us to
attach more metadata onto these StringMap types that we cannot encode in just a
string.

* openapi3: process discriminator mapping values as refs

While the type of the discriminator mapping values is a string in the
upstream specs, it contains a jsonschema reference to a schema object.
It is surprising behaviour that these refs are not handled when calling
functions such as InternalizeRefs.

This patch adds the data structures to store the ref internally, and
updates the Loader and InternalizeRefs to handle this case. There may be
several more functions that need to be updated that I am not aware of.

Since it is not a full Ref object we have to do some fudging to make it
work with all the existing ref handling code.