Fix RFC3339 validation
OpenAPI specifies the format `date-time` as RFC3339 compliant. In
a RFC3339 `date-time` the `time-offset` is a required component.

The current RFC3339 validation regex defines the `time-offset` as an
optional component. This leads to errors where a timestamp is considered
valid by the library but the value still cannot be parsed into a `time.Time`.

This change makes the `time-offset` a required component. Relevant tests are
changed accordingly.

Links:
- https://datatracker.ietf.org/doc/html/rfc3339#section-5.6

Fixes: #1031